Rename `EffectBundle` to `Effect`. It is now only used during related spawning.
